app/requirements.txt
pydantic~=2.9.2
SQLAlchemy~=2.0.35
fastapi~=0.115.0
minio~=7.2.9
pillow~=10.4.0
starlette~=0.38.6
bcrypt==4.0.1
python-jose~=3.3.0
passlib~=1.7.4
psycopg2-binary~=2.9.9
python-multipart~=0.0.12
uvicorn~=0.31.0

